Updates api route names to be more consistent. Moves playData parsing file into the backend of the file structure.
