+++ server/scriptController.js
@@ -1,16 +1,14 @@
-const playData = require('../playData.js');
+const playData = require('./playData.js');
 const db = require('./models/actorModels');
 
 const scriptController = {
   getPlay: (req, res, next) => {
-    // TODO change reference to testplay or twelfthNightplay
     const { title } = req.params;
     res.locals.fullPlay = playData[title].fullPlay;
     return next();
   },
 
   getCharacterData: (req, res, next) => {
-    // TODO change reference to testplay or twelfthNightplay
     const { title } = req.params;
     res.locals.characterData = playData[title].characterObjs;
     return next();
