+++ server/server.js
@@ -22,25 +22,29 @@ app.get('/script/:title', scriptController.getPlay, (req, res) => {
 });
 
 // return the character objects
-app.get('/characterData/:title', scriptController.getCharacterData, (req, res) => {
-  return res.status(200).json(res.locals.characterData);
-});
-
-// get a list of all the actors
-app.get('/getActors', actorController.getActors, (req, res) => {
-  return res.status(200).json(res.locals.actorList);
-});
-
 app.get(
-  '/currentCharacters/:firstName/:lastName/:option',
+  '/characterData/:title',
+  scriptController.getCharacterData,
+  (req, res) => {
+    return res.status(200).json(res.locals.characterData);
+  }
+);
+// get a list of the actors assigned to the current actor
+app.get(
+  '/actors/:firstName/:lastName/:option',
   actorController.getActorCharacters,
   (req, res) => {
     return res.status(200).json(res.locals.currentCharactersList);
   }
 );
 
+// get a list of all the actors
+app.get('/actors', actorController.getActors, (req, res) => {
+  return res.status(200).json(res.locals.actorList);
+});
+
 // add a new actor to the db
-app.post('/newActor', actorController.newActor, (req, res) => {
+app.post('/actors', actorController.newActor, (req, res) => {
   return res.sendStatus(204);
 });
